requirements.txt
aiogram[fastapi]~=3.3.0
aiohttp~=3.9.5
python-dotenv~=1.0.1
SQLAlchemy~=2.0.30
asyncpg~=0.29.0